+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## v1.4.0 (not released yet)
+## v1.4.0 (not released yet, https://github.com/go-reform/reform/milestones/v1.4.0)
 
 * Go 1.10+ is now required.
+* Added [`Querier.Count`](https://godoc.org/gopkg.in/reform.v1#Querier.Count).
+  Thanks to [Simon Kamenetskiy](https://github.com/skamenetskiy).
 * Added support for [github.com/jackc/pgx](https://github.com/jackc/pgx) driver.
 
 ## v1.3.3 (2018-12-11, https://github.com/go-reform/reform/milestones/v1.3.3)